Move logic into common file
