+++ app/forms/firewall-rules-edit.tsx
@@ -29,7 +29,12 @@ import { ALL_ISH } from '~/util/consts'
 import { invariant } from '~/util/invariant'
 import { pb } from '~/util/path-builder'
 
-import { CommonFields } from './firewall-rules-common'
+import {
+  CommonFields,
+  defaultActiveSubforms,
+  submitDisabledMessage,
+  useSubformStates,
+} from './firewall-rules-common'
 import { valuesToRuleUpdate, type FirewallRuleValues } from './firewall-rules-util'
 
 EditFirewallRuleForm.loader = async ({ params }: LoaderFunctionArgs) => {
@@ -54,6 +59,8 @@ export function EditFirewallRuleForm() {
   const { project, vpc, rule } = useFirewallRuleSelector()
   const vpcSelector = useVpcSelector()
   const queryClient = useApiQueryClient()
+
+  const { subformStates, updateSubformStates } = useSubformStates(defaultActiveSubforms)
 
   const { data: firewallRules } = usePrefetchedApiQuery('vpcFirewallRulesView', {
     query: { project, vpc },
@@ -122,6 +129,7 @@ export function EditFirewallRuleForm() {
       // validationSchema={validationSchema}
       // validateOnBlur
       loading={updateRules.isPending}
+      submitDisabled={submitDisabledMessage(subformStates)}
       submitError={updateRules.error}
     >
       <CommonFields
@@ -129,6 +137,7 @@ export function EditFirewallRuleForm() {
         // error if name is being changed to something that conflicts with some other rule
         nameTaken={(name) => !!otherRules.find((r) => r.name === name)}
         error={updateRules.error}
+        updateSubformStates={updateSubformStates}
       />
     </SideModalForm>
   )
